Update uvotLightCurveScript.py
There was a mistake in the previous commit where the Fermi data only plotted for a single time due and also the name of the fits file where the plot for the different filter got its data was incorrect. Both errors have been fixed. The final plot has also been changed to fit the output figure screen, the title has been moved to the top and the legend has been readded.